Refactoring code after merge.
