--- client/src/app/user.ts
@@ -1,8 +1,0 @@
-export class User {
-    userID: number;
-    username: string;
-    password: string;
-    token: string;
-    loggedIn: boolean;
-    message: string;
-}
